MS3

+++ backend/api/routes/poll.routes.js
@@ -1,30 +1,30 @@
 // Sends some sample data to mongodb server
 const mongoose = require('mongoose');
-require('dotenv').config();
 const User = require("../../models/user.js");
 const Poll = require("../../models/poll.js")
 const express = require("express");
 const router = express.Router();
 const { checkSession, checkCreateValidPoll } = require('../middleware.js')
 
-// Connect to mongodb
-mongoose.connect(process.env.CONNECTION_STRING, {
-    useNewUrlParser: true,
-    useUnifiedTopology: true,
-}).then(() => console.log('Connected to MongoDB'))
-    .catch(err => console.error('Could not connect to MongoDB...', err));
-
 //Record a vote from a user. If sucessful, appends _id to  the created_poll_id and answered_poll_id of this user. 
 //Also updates the responses of the poll with the request body.
 router.patch('/:id/vote', checkSession, async (req, res) => {
     const _id = req.params.id;
+    const answer = req.body.answer;
     try {
+        if (!mongoose.Types.ObjectId.isValid(_id)) {
+            return res.status(400).json({ message: 'Invalid ID format.' });
+        }
+        if(!Number.isInteger(answer)) {
+            return res.status(400).json({ message: 'Answer must be an integer.' });
+        }
+
         // Check if the poll is available directly
         const poll = await Poll.findById(_id);
         if (!poll) {
-            return res.status(400).json({ message: "Poll is not available" });
+            return res.status(403).json({ message: "Poll does not exist." });
         } else if (!poll.available) {
-            return res.status(400).json({ message: "Poll is not accepting responses" });
+            return res.status(403).json({ message: "Poll is not accepting responses." });
         }
 
         // Update user's answered_poll_id without adding duplicates (like adding to a set)
@@ -38,13 +38,13 @@ router.patch('/:id/vote', checkSession, async (req, res) => {
 
         if (existingResponse) {
             // Update the existing response
-            existingResponse.answer = req.body.answer;
+            existingResponse.answer = answer;
             existingResponse.updatedAt = Date.now();
         } else {
             // Add new response
             poll.responses.push({
                 user: req.session.userId,
-                answer: req.body.answer,
+                answer: answer,
                 updatedAt: Date.now()
             });
         }
@@ -53,7 +53,13 @@ router.patch('/:id/vote', checkSession, async (req, res) => {
         res.send(newPoll);
 
     } catch (error) {
-        res.status(500).json({ message: error.message });
+        if (error.name === 'ValidationError') {
+            const messages = Object.values(error.errors).map(error => error.message);
+            res.status(400).json({ errors: messages });
+        } else {
+            console.error('Error voting on poll:', error);
+            res.status(500).send();
+        }
     }
 });
 
@@ -62,11 +68,14 @@ router.patch('/:id/available', checkSession, async (req, res) => {
     const _id = req.params.id;
     const { available } = req.body;
 
-    if (typeof available !== 'boolean') {
-        return res.status(400).send({ message: "'available' must be true or false." });
-    }    
-
     try {
+        if (!mongoose.Types.ObjectId.isValid(_id)) {
+            return res.status(400).json({ message: 'Invalid ID format' });
+        }
+        if (typeof available !== 'boolean') {
+            return res.status(400).json({ message: "'available' must be true or false." });
+        }
+
         const poll = await Poll.findById(_id);
         if (req.session.userId !== poll.created_by.toString()) {
             return res.status(403).send("Forbidden");
@@ -78,11 +87,12 @@ router.patch('/:id/available', checkSession, async (req, res) => {
             res.send(updatedPoll);
         } else {
             // Do nothing if the current state matches the requested state
-            res.send({ message: "No changes made, poll availability is already set to " + available });
+            res.json({ message: "No changes made, poll availability is already set to " + available });
         }
     }
     catch (error) {
-        res.status(500).send({ message: error.message });
+        console.error('Error changing poll availability:', error);
+        res.status(500).send();
     }
 })
 
@@ -112,10 +122,16 @@ router.post('/', checkSession, checkCreateValidPoll, async (req, res) => {
             { $push: { created_poll_id: newPoll._id } }
         );
 
-        res.json(newPoll);
+        res.status(201).json(newPoll);
     }
     catch (error) {
-        res.status(500).send({ message: error.message });
+        if (error.name === 'ValidationError') {
+            const messages = Object.values(error.errors).map(error => error.message);
+            res.status(400).json({ errors: messages });
+        } else {
+            console.error('Error creating poll:', error);
+            res.status(500).send();
+        }
     }
 });
 
@@ -146,13 +162,14 @@ router.get('/:id', checkSession, async (req, res) => {
         // Conditionally modify the poll object based on who is requesting
         if (req.session.userId !== poll.created_by.toString()) {
             // If the user is not the creator of the poll, delete the correct answer and filter the responses
-            delete poll.correct_answer;
+            delete poll.correct_option;
             poll.responses = poll.responses.filter(response => response.user.toString() === req.session.userId);
         }
 
         res.send(poll);
     } catch (error) {
-        res.status(500).send({ message: error.message });
+        console.error('Error retrieving poll:', error);
+        res.status(500).send();
     }
 });
 
@@ -160,17 +177,19 @@ router.get('/:id', checkSession, async (req, res) => {
 router.delete('/:id', checkSession, async (req, res) => {
     try {
         const _id = req.params.id;
+        if (!mongoose.Types.ObjectId.isValid(_id)) {
+            return res.status(400).send({ message: 'Invalid ID format' });
+        }
         const poll = await Poll.findById(_id).select('created_by');
         if (!poll) {
             return res.status(404).send({ message: "Can't delete poll: Poll not found" });
         }
         if (req.session.userId !== poll.created_by.toString()) {
-            return res.status(403).send({ message: "Can't delete poll: Unauthorized" });
+            return res.status(403).send({ message: "Can't delete poll: Forbidden" });
         }
 
         // Poll Side
         await Poll.findOneAndDelete({ _id: _id })
-
 
         // User side 
         // Each poll is created by ONLY 1 user
@@ -190,20 +209,23 @@ router.delete('/:id', checkSession, async (req, res) => {
         res.send({ message: "Poll and references deleted successfully.", poll });
 
     } catch (error) {
-        // If there's an error, it might be because the `id` is not a valid ObjectId
-        res.status(500).send({ message: error.message });
+        console.error('Error deleting poll:', error);
+        res.status(500).send();
     }
 })
 
 router.patch('/:id/clear', checkSession, async (req, res) => {
     try {
         const _id = req.params.id;
+        if (!mongoose.Types.ObjectId.isValid(_id)) {
+            return res.status(400).send({ message: 'Invalid ID format' });
+        }
         const poll = await Poll.findById(_id).select('created_by');
         if (!poll) {
             return res.status(404).send({ message: "Can't clear poll: Poll not found" });
         }
         if (req.session.userId !== poll.created_by.toString()) {
-            return res.status(403).send({ message: "Can't clear poll: Unauthorized" });
+            return res.status(403).send({ message: "Can't clear poll: Forbidden." });
         }
         const result = await Poll.updateOne(
             { _id: _id },
@@ -224,8 +246,7 @@ router.patch('/:id/clear', checkSession, async (req, res) => {
 
     } catch (error) {
         console.error('Error clearing poll responses:', error);
-        // If there's an error, it might be because the `id` is not a valid ObjectId
-        res.status(500).send({ message: error.message });
+        res.status(500).send();
     }
 })
